# api/openapi.yaml
openapi: 3.0.3
info:
  title: FinishLine API
  description: Backend for footrace participant registration.
  version: 0.7.0
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
  /api/v1/registrations:
    post:
      operationId: register
      summary: Register for a race (machine-to-machine, called by the frontend BFF)
      description: >-
        Not a public browser endpoint — the only caller is our own Astro BFF,
        authenticated by a shared secret in the `X-Service-Secret` header
        (constant-time compared, same pattern as the Sanity webhook's
        `X-Webhook-Secret`). Upserts the person by email (one participant per
        email) and creates their registration for the race. Free races are
        confirmed immediately, assigning a sequential dorsal and sending a
        confirmation email.
      tags:
        - registrations
      parameters:
        - name: X-Service-Secret
          in: header
          required: true
          description: Shared BFF↔Go secret (`SERVICE_SECRET` on this API, `BACKEND_SERVICE_SECRET` on the BFF)
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Registered (and confirmed, for free races)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistrationResponse"
        "400":
          description: Malformed body, invalid race id, bad birth_date, or a form validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: birth date cannot be in the future
        "401":
          description: Missing or wrong X-Service-Secret
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: invalid service secret
        "404":
          description: The referenced race does not exist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: race not found
        "409":
          description: Already registered for this race, or the race is full
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: already registered for this race
    get:
      operationId: listRegistrations
      summary: Admin report — a race's registrations with dorsals
      tags:
        - registrations
      security:
        - bearerAuth: []
      parameters:
        - name: race_id
          in: query
          required: true
          description: The race to report on, by our internal race_id
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Registrations for the race, ordered by dorsal
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RegistrationReportRow"
        "400":
          description: Missing or invalid race_id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/webhooks/sanity:
    post:
      operationId: sanityWebhook
      summary: Sync a race from Sanity
      description: >-
        Machine-to-machine endpoint called by a Sanity GROQ-powered webhook on
        race document changes. Authenticated by a shared secret in the
        `X-Webhook-Secret` header (not a user token). `create`/`update`
        operations upsert the local race reference; `delete` removes it.
        Unknown/unmapped operations are acknowledged and ignored. This
        replaces the former `/webhooks/strapi` endpoint — Strapi is fully
        decommissioned as a CMS.
      tags:
        - webhooks
      parameters:
        - name: X-Webhook-Secret
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SanityWebhook"
      responses:
        "200":
          description: Event processed (synced, removed, or ignored)
        "400":
          description: Invalid payload or missing slug
        "401":
          description: Missing or wrong webhook secret
  /api/v1/races:
    get:
      operationId: listRaces
      summary: List registrable races (admin)
      description: >-
        Admin-only. The panel lists the races we mirror to obtain the
        `race_id` it needs to pull the registrations report. Each row carries
        our internal `race_id` and the Sanity slug as `document_id`.
        Registration itself uses the `document_id`, so the public form never
        calls this.
      tags:
        - races
      security:
        - bearerAuth: []
      responses:
        "200":
          description: The registrable races, ordered by date
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Race"
        "401":
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Race:
      type: object
      properties:
        race_id:
          type: string
          format: uuid
          description: Our internal id — use it to register and to pull the report
        document_id:
          type: string
          description: The Sanity slug — use it to fetch display content from Sanity
          example: renacer
        name:
          type: string
          example: Carrera 10K
        date:
          type: string
          format: date
          example: "2026-08-15"
        capacity:
          type: integer
          example: 500
    RegisterRequest:
      type: object
      required:
        - race_document_id
        - first_names
        - last_names
        - email
        - phone
        - document_id
        - birth_date
        - gender
        - referral_source
      properties:
        race_document_id:
          type: string
          description: The race's Sanity slug (Sanity owns the registration content, not the registration flow)
          example: renacer
        first_names:
          type: string
          example: Amir
        last_names:
          type: string
          example: Rojas
        email:
          type: string
          format: email
          example: amir@example.com
        phone:
          type: string
          example: "+59171234567"
        document_id:
          type: string
          description: The runner's CI (cédula de identidad). Stored only — not validated against RENIEC/SEGIP, not enforced unique.
          example: "1234567"
        birth_date:
          type: string
          format: date
          description: ISO date, YYYY-MM-DD
          example: "2000-06-09"
        gender:
          type: string
          enum: [M, F, X]
        referral_source:
          type: string
          example: Instagram
        modalidad:
          type: string
          description: The distance/variant chosen on the detail page (e.g. "10K · Con polera"). Optional — display data, not validated.
          example: "10K · Con polera"
    RegistrationResponse:
      type: object
      properties:
        registration_id:
          type: string
          format: uuid
        race_document_id:
          type: string
          description: The race's Sanity slug (echoes the request)
        first_names:
          type: string
        last_names:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        document_id:
          type: string
          description: The runner's CI (cédula de identidad)
        gender:
          type: string
          enum: [M, F, X]
        modalidad:
          type: string
          description: The distance/variant chosen on the detail page; empty when not provided
        status:
          type: string
          enum: [pendiente, confirmado]
        dorsal:
          type: integer
          nullable: true
          description: Assigned on confirmation; null while pending
        created_at:
          type: string
          format: date-time
        confirmed_at:
          type: string
          format: date-time
          nullable: true
    RegistrationReportRow:
      type: object
      properties:
        registration_id:
          type: string
          format: uuid
        first_names:
          type: string
        last_names:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        gender:
          type: string
          enum: [M, F, X]
        status:
          type: string
          enum: [pendiente, confirmado]
        dorsal:
          type: integer
          nullable: true
        created_at:
          type: string
          format: date-time
        confirmed_at:
          type: string
          format: date-time
          nullable: true
    SanityWebhook:
      type: object
      description: >-
        Shaped by the GROQ projection configured in the Studio webhook UI;
        NOT yet verified against the real webhook config (see design doc,
        open question) — flat race fields plus `operation` from Sanity's
        `delta::operation()` GROQ helper.
      required:
        - operation
      properties:
        operation:
          type: string
          enum: [create, update, delete]
          example: create
        slug:
          type: string
          description: The race's Sanity slug (external id)
          example: renacer
        title:
          type: string
          example: Carrera 10K
        date:
          type: string
          format: date
          example: "2026-08-15"
        capacity:
          type: integer
          example: 500
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: email is already registered